+++ src/index.ts
@@ -19,3 +19,6 @@ export type {
     Overrides,
     FieldType,
 } from './builder/types';
+
+export {resetable} from 'reset/resetable';
+export type {ResetSignal} from 'reset/ResetSignal';